pyproject.toml
[project]
name = "agglovar"
description = "Toolkit for fast genomic variant transformations and intersects"

license = "MIT"
license-files = ["LICENSE"]

dynamic = ["version"]

readme = "README.md"

requires-python = ">=3.12"
dependencies = [
    "polars>=1.33.1",
    "biopython>=1.85",
    "edlib>=1.3.9.post1",
    "numpy>=2.3.0",
    "ply>=3.11",
]

[agglovar.urls]
Homepage = "https://github.com/BeckLaboratory/agglovar"
#Documentation = "https://agglovar.readthedocs.io"
Repository = "https://github.com/BeckLaboratory/agglovar"
Issues = "https://github.com/BeckLaboratory/agglovar/issues"
Changelog = "https://github.com/BeckLaboratory/agglovar/blob/master/CHANGELOG.md"

keywords = [
    "variant merging",
    "variant intersect",
    "bioinformatics",
    "genomics"
]

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Private :: Do Not Upload",  # Temporary until published
]

authors = [
    {name = "Peter Audano", email = "peter.audano@jax.org"},
]

maintainers = [
    {name = "Peter Audano", email = "peter.audano@jax.org"}
]

[dependency-groups]
dev = [
    "flake8>=7.3.0",
    "flake8-bugbear>=24.12.12",
    "flake8-docstrings>=1.7.0",
    "flake8-dunder-all>=0.5.0",
    "flake8-mutable>=1.2.0",
    "flake8-pyproject>=1.2.3",
    "flake8-rst-docstrings>=0.3.1",
    "ipython>=9.4.0",
    "pylint>=3.3.7",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "sphinx>=8.2.3",
    "sphinx-autoapi>=3.6.0",
    "sphinx-rtd-theme>=3.0.2",
    "tox>=4.30.2",
    "twine>=6.2.0",
]

[project.optional-dependencies]
plot = [
    "matplotlib>=3.10.3",
]
all = [
    "agglovar[plot]"
]


[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["agglovar"]
exclude = ["*__pycache__"]

[tool.setuptools.dynamic]
version = {attr = "agglovar.__version__"}


[tool.pytest.ini_options]
testpaths = [
    "tests"
]


[tool.pylint.'MESSAGES CONTROL']
max-line-length = 120

[tool.flake8]
max-line-length = 120
docstring-convention = "google"
per-file-ignores = [
    "src/agglovar/join/config/parser.py: D200, D205, D212, D300, D403, D415"
]
rst-roles = ["class", "meth", "const", "ref", "func"]
